Change subword iterator from using 32-bit words to 8-bit words

+++ tools/gfxc/subword_output_iterator.hpp
@@ -8,12 +8,12 @@ class subword_output_iterator_deref;
 // : public std::iterator<std::output_iterator_tag, void, void, void, void>
 class subword_output_iterator {
 private:
-	std::shared_ptr<std::vector<uint32_t>> _backing;
+	std::shared_ptr<std::vector<uint8_t>> _backing;
 	unsigned _word;
 	unsigned _subword;
 	unsigned _width;
 	subword_output_iterator(
-		std::shared_ptr<std::vector<uint32_t>> backing,
+		std::shared_ptr<std::vector<uint8_t>> backing,
 		unsigned word,
 		unsigned subword,
 		unsigned width);
@@ -26,17 +26,17 @@ public:
 	subword_output_iterator& operator++();
 	subword_output_iterator operator++(int);
 
-	std::vector<uint32_t> result();
+	std::vector<uint8_t> result();
 };
 
 class subword_output_iterator_deref {
 private:
-	std::shared_ptr<std::vector<uint32_t>> _backing;
+	std::shared_ptr<std::vector<uint8_t>> _backing;
 	unsigned _word;
 	unsigned _subword;
 	unsigned _width;
 	subword_output_iterator_deref(
-		std::shared_ptr<std::vector<uint32_t>> backing,
+		std::shared_ptr<std::vector<uint8_t>> backing,
 		unsigned word,
 		unsigned subword,
 		unsigned width);
